refact: KEYP-371 Mark SAM manager classes as deprecated and move them to a separate package
